Improved artifact, started Attribute items and removed prison stuff

+++ src/main/kotlin/com/cludivers/kz2wdprison/framework/persistance/beans/artifact/inputs/ArtifactInput.kt
@@ -6,7 +6,7 @@ import org.bukkit.util.Vector
 
 
 class ArtifactInput(
-    var flow: Int = 0,
+    var flow: Float = 0f,
     var locations: List<Location> = listOf(),
     var entities: List<Entity> = listOf(),
     var vectors: List<Vector> = listOf(),
